libxdiff: Fix types on adler32, the worst 32-bit hash.

+++ libxdiff/xdiff/xbdiff.h
@@ -28,7 +28,7 @@
 #define XDL_INSBOP_SIZE (1 + 4)
 #define XDL_COPYOP_SIZE (1 + 4 + 4)
 
-unsigned long xdl_mmb_adler32(mmbuffer_t *mmb);
-unsigned long xdl_mmf_adler32(mmfile_t *mmf);
+uint32_t xdl_mmb_adler32(mmbuffer_t *mmb);
+uint32_t xdl_mmf_adler32(mmfile_t *mmf);
 
 #endif /* #if !defined(XBDIFF_H) */
